Backend/requirements.txt
altgraph==0.17.2
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
Babel==2.13.1
bcrypt==4.0.1
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cycler==0.12.1
Flask==2.2.5
Flask-BabelEx==0.9.4
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==3.0.0
flask_security
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.45.1
future==0.18.2
greenlet==3.0.1
httplib2==0.22.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.4
macholib==1.15.2
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.41
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.15.0
speaklater==1.3
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==2.2.2
WTForms==3.1.1
zipp==3.17.0